% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Title}
\usage{
minus(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
Difference of 'x' and 'y'.
}
\description{
This function returns a difference between two numbers.
}
\examples{
minus(3,2)
}
