% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Plus}
\usage{
plus(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
Sum of 'x' and 'y'.
}
\description{
This function returns a sum of two numbers.
}
\examples{
plus(2,3)
}
